+++ spring-cloud-gcp-data-spanner/src/test/java/org/springframework/cloud/spanner/core/mapping/SpannerMappingContextTests.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package org.springframework.data.spanner.core.mapping;
+package org.springframework.cloud.spanner.core.mapping;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
