+++ spring-cloud-gcp-data-spanner/src/test/java/org/springframework/cloud/spanner/core/mapping/SpannerPersistentEntityImplTests.java
@@ -14,7 +14,7 @@
  *  limitations under the License.
  */
 
-package org.springframework.data.spanner.core.mapping;
+package org.springframework.cloud.spanner.core.mapping;
 
 import org.junit.Test;
 
@@ -63,12 +63,12 @@ public class SpannerPersistentEntityImplTests {
 		assertThat(entity.columns(), containsInAnyOrder("custom_col", "id"));
 	}
 
-	@Table(name = "custom_test_table")
+	@SpannerTable(name = "custom_test_table")
 	private static class TestEntity {
 		@Id
 		String id;
 
-		@Column(name = "custom_col")
+		@SpannerColumn(name = "custom_col")
 		String something;
 	}
 
@@ -79,7 +79,7 @@ public class SpannerPersistentEntityImplTests {
 		String something;
 	}
 
-	@Table(name = "")
+	@SpannerTable(name = "")
 	private static class EntityEmptyCustomName {
 		@Id
 		String id;
